EPR-28: crecion y configuracion de archivos

+++ demo/src/main/java/com/example/demo/repository/NeighborhoodRepository.java
@@ -1,5 +1,5 @@
 package com.example.demo.repository;
 import com.example.demo.model.Neighborhood;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface NeighborhoodRepository extends JpaRepository<Neighborhood, Integer> { }
+public interface NeighborhoodRepository extends JpaRepository<Neighborhood, Long> { }
